requirements.txt
asgiref==3.3.1
dataclasses==0.8
Django==3.1.5
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
typing-extensions==3.7.4.3
